# functions/lib/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FeedSolve API",
        "version": "1.0.0",
        "description": "REST API for feedback management and analytics",
        "contact": {
            "name": "FeedSolve Support",
            "url": "https://feedsolve.com/support"
        }
    },
    "servers": [
        {
            "url": "https://api.feedsolve.com",
            "description": "Production API"
        },
        {
            "url": "http://localhost:5001",
            "description": "Local development"
        }
    ],
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "paths": {
...
        "/api/boards": {
            "post": {
                "summary": "Create a new board",
                "tags": ["Boards"],
                "operationId": "createBoard",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BoardInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Board created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/boards/{id}": {
            "get": {
                "summary": "Get board by ID",
                "tags": ["Boards"],
                "operationId": "getBoard",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Board found"
                    },
                    "404": {
                        "description": "Board not found"
                    }
                }
            },
            "patch": {
                "summary": "Update board",
                "tags": ["Boards"],
                "operationId": "updateBoard",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BoardInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Board updated"
                    },
                    "404": {
                        "description": "Board not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete board",
                "tags": ["Boards"],
                "operationId": "deleteBoard",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Board deleted"
                    },
                    "404": {
                        "description": "Board not found"
                    }
                }
            }
        },
...
        "/api/auth/api-keys": {
            "get": {
                "summary": "List API keys",
                "tags": ["Authentication"],
                "operationId": "listApiKeys",
                "responses": {
                    "200": {
                        "description": "List of API keys"
                    }
                }
            },
            "post": {
                "summary": "Create API key",
                "tags": ["Authentication"],
                "operationId": "createApiKey",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ApiKeyInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "API key created (key shown only once)"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/auth/api-keys/{keyId}": {
            "delete": {
                "summary": "Delete API key",
                "tags": ["Authentication"],
                "operationId": "deleteApiKey",
                "parameters": [
                    {
                        "name": "keyId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "API key deleted"
                    },
                    "404": {
                        "description": "API key not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BoardInput": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "isAnonymousAllowed": {
                        "type": "boolean"
                    }
                },
                "required": ["name"]
            },
...
            "ApiKeyInput": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "submissions:read",
                                "submissions:write",
                                "submissions:delete",
                                "boards:read",
                                "boards:write",
                                "boards:delete",
                                "stats:read",
                                "keys:create",
                                "keys:read",
                                "keys:delete",
                                "company:read"
                            ]
                        }
                    },
                    "expiresAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "ipWhitelist": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "required": ["name", "permissions"]
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "token",
                "description": "API Key in format: Bearer fsk_xxxxx"
            }
        }
    }
}